Move Accumulator and GroupedAccumulator to spi
