+++ presto-main/src/main/java/com/facebook/presto/GroupByHashPageIndexer.java
@@ -14,10 +14,10 @@
 package com.facebook.presto;
 
 import com.facebook.presto.operator.GroupByHash;
-import com.facebook.presto.operator.GroupByIdBlock;
 import com.facebook.presto.operator.Work;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.PageIndexer;
+import com.facebook.presto.spi.block.GroupByIdBlock;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.sql.gen.JoinCompiler;
