+++ presto-main/src/main/java/com/facebook/presto/operator/aggregation/GenericAccumulatorFactoryBinder.java
@@ -16,8 +16,10 @@ package com.facebook.presto.operator.aggregation;
 import com.facebook.presto.Session;
 import com.facebook.presto.operator.PagesIndex;
 import com.facebook.presto.spi.block.SortOrder;
+import com.facebook.presto.spi.function.Accumulator;
 import com.facebook.presto.spi.function.AccumulatorStateFactory;
 import com.facebook.presto.spi.function.AccumulatorStateSerializer;
+import com.facebook.presto.spi.function.GroupedAccumulator;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.sql.gen.JoinCompiler;
 import com.google.common.annotations.VisibleForTesting;
