+++ presto-main/src/test/java/com/facebook/presto/operator/aggregation/groupByAggregations/AggregationTestOutput.java
@@ -15,8 +15,8 @@
 package com.facebook.presto.operator.aggregation.groupByAggregations;
 
 import com.facebook.presto.block.BlockAssertions;
-import com.facebook.presto.operator.aggregation.GroupedAccumulator;
 import com.facebook.presto.spi.block.BlockBuilder;
+import com.facebook.presto.spi.function.GroupedAccumulator;
 
 import java.util.function.BiConsumer;
