+++ presto-spi/src/main/java/com/facebook/presto/spi/function/Accumulator.java
@@ -11,12 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.operator.aggregation;
+package com.facebook.presto.spi.function;
 
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
-import com.facebook.presto.spi.function.WindowIndex;
 import com.facebook.presto.spi.type.Type;
 
 import java.util.List;
